+++ baidu-netdisk/src/main/kotlin/xyz/cssxsh/baidu/disk/item/NetDiskDoc.kt
@@ -37,7 +37,7 @@ data class NetDiskDoc(
     @SerialName("share")
     val share: Int,
     @SerialName("size")
-    override val size: Int,
+    override val size: Long,
     @SerialName("thumbs")
     val thumbs: Map<String, String>? = null,
     @SerialName("wpfile")
